# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    ContentItem:
      type: object
      required: [id, type, title, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [Film, TVSeries, TBN_FAST, TBN_Linear, WoF_FAST]
        title:
          type: string
        description:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/Season'
        contractCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Season:
      type: object
      required: [id, contentItemId, seasonNumber]
      properties:
        id:
          type: string
        contentItemId:
          type: string
        seasonNumber:
          type: integer
        title:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        episodeCount:
          type: integer
          nullable: true
...
    Contract:
      type: object
      required: [id, direction, partnerId, status, endType, createdAt]
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        partnerName:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        otherTerritories:
          type: string
          nullable: true
        distributionTypes:
          type: array
          items:
            type: string
        platform:
          type: string
          nullable: true
        royaltyType:
          type: string
          enum: [revenue_share, flat_fee, other]
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        paymentTerms:
          type: string
          enum: [net_30, net_60, net_90]
          nullable: true
        notes:
          type: string
          nullable: true
        documentUrl:
          type: string
          nullable: true
        websiteLink:
          type: string
          nullable: true
        rightsInDetails:
          $ref: '#/components/schemas/RightsInDetails'
          nullable: true
        rightsOutDetails:
          $ref: '#/components/schemas/RightsOutDetails'
          nullable: true
        contentItems:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        amendments:
          type: array
          items:
            $ref: '#/components/schemas/Amendment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdByName:
          type: string
          nullable: true
    RightsInDetails:
      type: object
      properties:
        platforms:
          type: array
          items:
            type: string
        youtubeChannel:
          type: string
          nullable: true
        socialPlatforms:
          type: array
          items:
            type: string
        socialHandle:
          type: string
          nullable: true
        grantOfRights:
          type: string
          nullable: true
        exclusivityStartDate:
          type: string
          format: date
          nullable: true
        exclusivityEndDate:
          type: string
          format: date
          nullable: true
        exclusivitySameAsDuration:
          type: boolean
        marketingRights:
          type: string
          nullable: true
    RightsOutDetails:
      type: object
      properties:
        autoRenew:
          type: boolean
        hasAmendment:
          type: boolean
        exclusivity:
          type: string
          enum: [exclusive, non_exclusive]
          nullable: true
        reportingFrequency:
          type: string
          enum: [monthly, quarterly, annually]
          nullable: true
        minPaymentThreshold:
          type: number
          nullable: true
    CreateContractRequest:
      type: object
      required: [direction, partnerId, endType]
      properties:
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        otherTerritories:
          type: string
          nullable: true
        distributionTypes:
          type: array
          items:
            type: string
        platform:
          type: string
          nullable: true
        royaltyType:
          type: string
          enum: [revenue_share, flat_fee, other]
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        paymentTerms:
          type: string
          enum: [net_30, net_60, net_90]
          nullable: true
        notes:
          type: string
          nullable: true
        websiteLink:
          type: string
          nullable: true
        rightsInDetails:
          $ref: '#/components/schemas/RightsInDetails'
          nullable: true
        rightsOutDetails:
          $ref: '#/components/schemas/RightsOutDetails'
          nullable: true
        contentItemIds:
          type: array
          items:
            type: string
    UpdateContractRequest:
      type: object
      properties:
        partnerId:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        otherTerritories:
          type: string
          nullable: true
        distributionTypes:
          type: array
          items:
            type: string
        platform:
          type: string
          nullable: true
        royaltyType:
          type: string
          enum: [revenue_share, flat_fee, other]
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        paymentTerms:
          type: string
          enum: [net_30, net_60, net_90]
          nullable: true
        notes:
          type: string
          nullable: true
        websiteLink:
          type: string
          nullable: true
        rightsInDetails:
          $ref: '#/components/schemas/RightsInDetails'
          nullable: true
        rightsOutDetails:
          $ref: '#/components/schemas/RightsOutDetails'
          nullable: true
        contentItemIds:
          type: array
          items:
            type: string
    ContractListItem:
      type: object
      required: [id, direction, status, endType, createdAt]
      properties:
        id:
          type: string
        direction:
          type: string
          enum: [rights_in, rights_out]
        partnerId:
          type: string
        partnerName:
          type: string
        licensor:
          type: string
          nullable: true
        licensee:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, expired, in_perpetuity, terminated]
        startDate:
          type: string
          format: date
          nullable: true
        endType:
          type: string
          enum: [date, perpetuity, auto_renew]
        endDate:
          type: string
          format: date
          nullable: true
        territories:
          type: array
          items:
            type: string
        distributionTypes:
          type: array
          items:
            type: string
        royaltyType:
          type: string
          nullable: true
        contentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    Amendment:
      type: object
      required: [id, contractId, date, description, createdAt]
      properties:
        id:
          type: string
        contractId:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdByName:
          type: string
          nullable: true
    CreateAmendmentRequest:
      type: object
      required: [date, description]
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: string
          nullable: true
...
    ContractImportInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    ImportResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: object
            required: [row, message]
            properties:
              row:
                type: integer
              message:
                type: string
    PaginatedContracts:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContractListItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /contracts:
    get:
      operationId: listContracts
      summary: List contracts with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: direction
          in: query
          schema:
            type: string
            enum: [rights_in, rights_out]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, expired, in_perpetuity, terminated]
        - name: partnerId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: expiringWithinDays
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated contract list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedContracts'
    post:
      operationId: createContract
      summary: Create a new contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: Created contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{id}:
    get:
      operationId: getContract
      summary: Get contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contract details with content and amendments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateContract
      summary: Update contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
    delete:
      operationId: deleteContract
      summary: Delete contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /contracts/{id}/amendments:
    get:
      operationId: listAmendments
      summary: List amendments for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of amendments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Amendment'
    post:
      operationId: createAmendment
      summary: Add an amendment to a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAmendmentRequest'
      responses:
        '201':
          description: Created amendment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Amendment'
...
  /content/{id}/contracts:
    get:
      operationId: getContentContracts
      summary: Get all contracts linked to a content item (reverse lookup)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contracts linked to this content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContractListItem'
...
  /import/contracts:
    post:
      operationId: importContracts
      summary: Bulk import contracts from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ContractImportInput'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
...
  /reports/contracts:
    get:
      operationId: getContractReport
      summary: Get contracts report data
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [rights_in, rights_out]
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, xlsx, pdf]
            default: json
      responses:
        '200':
          description: Contract report (JSON or file)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractListItem'
                  generatedAt:
                    type: string
                    format: date-time
...