# lib/api-zod/src/generated/types/createContractRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContractRequestDirection } from './createContractRequestDirection';
import type { CreateContractRequestEndType } from './createContractRequestEndType';
import type { CreateContractRequestPaymentTerms } from './createContractRequestPaymentTerms';
import type { CreateContractRequestRoyaltyType } from './createContractRequestRoyaltyType';
import type { CreateContractRequestStatus } from './createContractRequestStatus';
import type { RightsInDetails } from './rightsInDetails';
import type { RightsOutDetails } from './rightsOutDetails';

export interface CreateContractRequest {
  direction: CreateContractRequestDirection;
  partnerId: string;
  licensor?: string | null;
  licensee?: string | null;
  status?: CreateContractRequestStatus;
  startDate?: Date | null;
  endType: CreateContractRequestEndType;
  endDate?: Date | null;
  territories?: string[];
  otherTerritories?: string | null;
  distributionTypes?: string[];
  platform?: string | null;
  royaltyType?: CreateContractRequestRoyaltyType;
  royaltyDetails?: string | null;
  paymentTerms?: CreateContractRequestPaymentTerms;
  notes?: string | null;
  websiteLink?: string | null;
  rightsInDetails?: RightsInDetails | null;
  rightsOutDetails?: RightsOutDetails | null;
  contentItemIds?: string[];
}